Handle logout in session management, not in application version
